refactor(github-action): decreased number of files and created 1 file
