--- test/github-action/config/git.test.ts
@@ -1,73 +1,0 @@
-/**
- * Test files test the whole results of the Projen project creation
- * configuration orchestration. E.g. the npm test module does test the result of the whole `package.json` file
- * and that the devcontainer config module scripts are created. The devcontainer test module itself will not test
- * if the related devcontainer npm scripts are created.
- *
- * The `setup.ts` file includes all relevant bootstrap steps for the test files.
- * It is important that each test file imports and uses the `snapshot` otherwise the bootstrap will not run for this test file.
- **/
-
-import * as git from '../../shared/git';
-import { snapshot } from './setup';
-
-/**
- * We use not a describe block here because the jest test runner `Test Results` pane will show the test names.
- * The `Testing` pane shows a hierarchy but for this hierarchy we already have the test file names available.
- */
-test('Ignore patterns matches expected content', (): void => {
-  const expectedEntries: string[] = [
-    '',
-    '!/.commitlintrc.ts',
-    '!/.devcontainer.json',
-    '!/.gitattributes',
-    '!/.github/ISSUE_TEMPLATE/bug.yml',
-    '!/.github/ISSUE_TEMPLATE/feature.yml',
-    '!/.github/ISSUE_TEMPLATE/housekeeping.yml',
-    '!/.github/ISSUE_TEMPLATE/question.yml',
-    '!/.github/pull_request_template.md',
-    '!/.github/workflows/enforce-labels.yml',
-    '!/.github/workflows/release.yml',
-    '!/.husky/commit-msg',
-    '!/.husky/pre-commit',
-    '!/.prettierignore',
-    '!/.prettierrc.json',
-    '!/.projen/deps.json',
-    '!/.projen/files.json',
-    '!/.projen/tasks.json',
-    '!/.projenrc.js',
-    '!/.vscode/settings.json',
-    '!/cliff.toml',
-    '!/package.json',
-    '!/src/',
-    '!/test/',
-    '!/tsconfig.json',
-    '# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".',
-    '*.lcov',
-    '*.log',
-    '*.pid',
-    '*.pid.lock',
-    '*.seed',
-    '*.tgz',
-    '*.tsbuildinfo',
-    '.cache',
-    '.eslintcache',
-    '.nyc_output',
-    '.yarn-integrity',
-    '/dist/',
-    '/lib',
-    'build/Release',
-    'coverage',
-    'jspm_packages/',
-    'lerna-debug.log*',
-    'lib-cov',
-    'logs',
-    'node_modules/',
-    'npm-debug.log*',
-    'pids',
-    'report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json',
-    'yarn-debug.log*',
-    'yarn-error.log*',
-  ];
-  git.testIgnore(snapshot, expectedEntries);
-});
